# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Service Support API
  version: 1.0.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: [ ]
paths:
...
  /tools:
    get:
      tags: [ tool ]
      summary: Get all tools
      responses:
        '200':
          description: List of tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tool'
    post:
      tags: [ tool ]
      summary: Create new tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tool'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
  /tools/{id}:
    get:
      tags: [ tool ]
      summary: Get tool by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tool found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
        '404':
          description: Not found
    put:
      tags: [ tool ]
      summary: Update tool
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tool'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
    delete:
      tags: [ tool ]
      summary: Delete tool
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Tool:
      type: object
      properties:
        id:
          type: integer
          example: 1
        tool_id:
          type: string
          example: 2025-10
        itemNumber:
          type: string
          example: "ITM-001"
        name:
          type: string
          example: "Fúrógép"
        serialNumber:
          type: string
          example: "SN-2025-0001"
        typeNumber:
          type: string
          example: "TYP-1234"
        ownerName:
          type: string
          description: Az ügyfél neve (owner_company_employee.name)
          example: "Kiss János"
        ownerCompanyName:
          type: string
          description: Az ügyfél cégének neve (ha van, különben üresen marad)
          example: "Szerszám Kuckó Kft."
...